+++ src/decorators/index.ts
@@ -1,3 +1,0 @@
-export * from './BuiltInstance';
-export * from './BuiltInstanceFactory';
-export * from './WorkerStatus';